refactor: Adjust folder naming structure
